% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_aggregate_daily.R
\name{aggregate2daily}
\alias{aggregate2daily}
\title{Aggregate data to daily.}
\usage{
aggregate2daily(dirAWS, netAWS, obs_hour = 8, archive = FALSE)
}
\arguments{
\item{dirAWS}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}

\item{netAWS}{the name of the AWS network, "LSI-ELOG", "LSI-XLOG" or "REMA".}

\item{obs_hour}{daily precipitation observation hour.}

\item{archive}{TRUE for archive mode.}
}
\description{
Aggregate AWS data to daily.
}
