% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_hourly.R
\name{aggregate2hourly}
\alias{aggregate2hourly}
\title{Aggregate data to hourly.}
\usage{
aggregate2hourly(dirAWS, netAWS, archive = FALSE)
}
\arguments{
\item{dirAWS}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}

\item{netAWS}{the name of the AWS network, "LSI-ELOG", "LSI-XLOG" or "REMA".}

\item{archive}{TRUE for archive mode.}
}
\description{
Aggregate AWS data to hourly.
}
