% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_check_variables.R
\name{aws_check_variables_ids}
\alias{aws_check_variables_ids}
\title{Checking AWS variables.}
\usage{
aws_check_variables_ids(start_time, end_time, aws_ids, variables, aws_dir)
}
\arguments{
\item{start_time}{start time, format "YYYY-mm-dd HH:MM"}

\item{end_time}{end time, format "YYYY-mm-dd HH:MM"}

\item{aws_ids}{aws ids, a vector of aws id
Example: c("301", "306", "10050037", "000003")}

\item{variables}{a vector of aws variables. Ex: c('RR', 'TT')}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get AWS variables for a list of AWS.
}
