% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_data_extraction.R
\name{aws_data_aggregate_ids}
\alias{aws_data_aggregate_ids}
\title{Get aggregated data.}
\usage{
aws_data_aggregate_ids(timestep, start_time, end_time, aws_ids, aws_dir)
}
\arguments{
\item{timestep}{time step of the data, "hourly", "daily", "pentad", "dekadal" or "monthly"}

\item{start_time}{start time, format "YYYY-mm-dd HH:00"}

\item{end_time}{end time, format "YYYY-mm-dd HH:00"}

\item{aws_ids}{aws ids, a vector of aws id.\cr
Example: c("301", "306", "10050037", "000003")}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get aggregated data for multiple AWS.
}
