% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_data_availability.R
\name{aws_data_availability_all}
\alias{aws_data_availability_all}
\title{AWS data availability.}
\usage{
aws_data_availability_all(start_time, end_time, aws_dir)
}
\arguments{
\item{start_time}{start time, format "YYYY-mm-dd HH:MM"}

\item{end_time}{end time, format "YYYY-mm-dd HH:MM"}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get AWS data availability for all networks 'LSI-XLOG', 'LSI-ELOG' and 'REMA'.
}
