% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_data_availability.R
\name{aws_data_availability_ids}
\alias{aws_data_availability_ids}
\title{AWS data availability.}
\usage{
aws_data_availability_ids(start_time, end_time, aws_ids, aws_dir)
}
\arguments{
\item{start_time}{start time, format "YYYY-mm-dd HH:MM"}

\item{end_time}{end time, format "YYYY-mm-dd HH:MM"}

\item{aws_ids}{aws ids, a vector of aws id
Example: c("301", "306", "10050037", "000003")}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get AWS data availability for a list of AWS.
}
