% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_data_availability.R
\name{aws_data_status_24hour}
\alias{aws_data_status_24hour}
\title{Get AWS Status data.}
\usage{
aws_data_status_24hour(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get AWS last 24 hour data availability.
}
