% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_data_extraction.R
\name{aws_hourly_precip_data}
\alias{aws_hourly_precip_data}
\title{Get hourly precipitation data.}
\usage{
aws_hourly_precip_data(start_time, end_time, aws_dir)
}
\arguments{
\item{start_time}{start time, format "YYYY-mm-dd HH:00"}

\item{end_time}{end time, format "YYYY-mm-dd HH:00"}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get hourly precipitation data for all AWS.
}
