% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_metadata.R
\name{aws_metadata}
\alias{aws_metadata}
\title{Get AWS metadat.}
\usage{
aws_metadata(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get AWS metadat.
}
