% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_spatial_daily.R
\name{aws_spatial_daily}
\alias{aws_spatial_daily}
\title{Create daily spatial data.}
\usage{
aws_spatial_daily(dirAWS)
}
\arguments{
\item{dirAWS}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\description{
Create daily spatial data.
}
