% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_spatial_daily.R
\name{aws_spatial_daily_arch}
\alias{aws_spatial_daily_arch}
\title{Create daily spatial data.}
\usage{
aws_spatial_daily_arch(start_date, end_date, dirAWS)
}
\arguments{
\item{start_date}{the start date to process in the format "YYYY-MM-DD".
Example: "2019-12-15"}

\item{end_date}{the end date to process in the format "YYYY-MM-DD"}

\item{dirAWS}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\description{
Create daily spatial data archive mode.
}
