% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_spatial_hourly.R
\name{aws_spatial_hourly}
\alias{aws_spatial_hourly}
\title{Create hourly spatial data.}
\usage{
aws_spatial_hourly(dirAWS)
}
\arguments{
\item{dirAWS}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\description{
Create hourly spatial data.
}
