% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_spatial_hourly.R
\name{aws_spatial_hourly_arch}
\alias{aws_spatial_hourly_arch}
\title{Create hourly spatial data.}
\usage{
aws_spatial_hourly_arch(start_time, end_time, dirAWS)
}
\arguments{
\item{start_time}{the start time to process in the format "YYYY-MM-DD HH:00".
Example: "2019-12-15 12:00"}

\item{end_time}{the end time to process in the format "YYYY-MM-DD HH:00"}

\item{dirAWS}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\description{
Create hourly spatial data archive mode.
}
