% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_spatial_minutes.R
\name{aws_spatial_minutes}
\alias{aws_spatial_minutes}
\title{Create 10 minutes spatial data.}
\usage{
aws_spatial_minutes(dirAWS)
}
\arguments{
\item{dirAWS}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\description{
Create 10 minutes spatial data.
}
