% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_spatial_minutes.R
\name{aws_spatial_minutes_arch}
\alias{aws_spatial_minutes_arch}
\title{Create 10 minutes spatial data.}
\usage{
aws_spatial_minutes_arch(start_time, end_time, dirAWS)
}
\arguments{
\item{start_time}{the start time to process in the format "YYYY-MM-DD HH:MM".
Example: "2019-12-15 12:50"}

\item{end_time}{the end time to process in the format "YYYY-MM-DD HH:MM"}

\item{dirAWS}{full path to the directory of the parsed data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\description{
Create 10 minutes spatial data archive mode.
}
