% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plotchart_aws.R
\name{chartAggrAWSData}
\alias{chartAggrAWSData}
\title{Get aggregated data.}
\usage{
chartAggrAWSData(
  tstep,
  aws,
  vars,
  pars,
  start,
  end,
  plotrange,
  aws_net,
  aws_dir
)
}
\arguments{
\item{tstep}{time step.}

\item{aws}{AWS ID.}

\item{vars}{variables.}

\item{pars}{parameters.}

\item{start}{start date.}

\item{end}{end date.}

\item{plotrange}{get range.}

\item{aws_net}{AWS network.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get aggregated data to display on chart.
}
