% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plotchart_aws.R
\name{chartMinAWSData}
\alias{chartMinAWSData}
\title{Get 10 minutes data.}
\usage{
chartMinAWSData(aws, vars, pars, start, end, plotrange, aws_net, aws_dir)
}
\arguments{
\item{aws}{AWS ID.}

\item{vars}{variables.}

\item{pars}{parameters.}

\item{start}{start time.}

\item{end}{end time.}

\item{plotrange}{get range.}

\item{aws_net}{AWS network.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get 10 minutes data to display on chart.
}
