% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rain_accumulation.R
\name{chartRainAccumulAWS}
\alias{chartRainAccumulAWS}
\title{Compute precipitation accumulation.}
\usage{
chartRainAccumulAWS(tstep, aws, start, end, accumul, aws_dir)
}
\arguments{
\item{tstep}{time basis to accumulate the data.}

\item{aws}{AWS ID.}

\item{start}{start date.}

\item{end}{end date.}

\item{accumul}{accumulation duration.}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Compute precipitation accumulation for chart display.
}
