% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_display_logfiles.R
\name{displayLogFiles}
\alias{displayLogFiles}
\title{Read log files.}
\usage{
displayLogFiles(logtype, aws_dir, aws_net = NULL, date = NULL)
}
\arguments{
\item{logtype}{type of log files: "logperaws", "logproc" or "logupload".}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}

\item{aws_net}{AWS network: "REMA", "LSI-XLOG" or "LSI-ELOG".}

\item{date}{date to read.}
}
\value{
JSON object
}
\description{
Read log files to be displayed.
}
