% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plotmap_aws.R
\name{displayMAP10min}
\alias{displayMAP10min}
\title{Get AWS 10 minutes spatial data.}
\usage{
displayMAP10min(time, aws_dir)
}
\arguments{
\item{time}{the time to display in the format "YYYY-MM-DD-HH-MM".}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get AWS 10 minutes spatial data to display on map.
}
