% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rain_accumulation.R
\name{displayMAPRainAccumul}
\alias{displayMAPRainAccumul}
\title{Compute precipitation accumulation.}
\usage{
displayMAPRainAccumul(tstep, time, accumul, aws_dir)
}
\arguments{
\item{tstep}{time basis to accumulate the data.}

\item{time}{target date.}

\item{accumul}{accumulation duration.}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Compute precipitation accumulation for spatial display.
}
