% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_display_output_qc.R
\name{displayQCMinutes}
\alias{displayQCMinutes}
\title{Get QC output for minutes data.}
\usage{
displayQCMinutes(time, aws_dir)
}
\arguments{
\item{time}{time range in hour to parse.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
JSON object
}
\description{
Get QC output for minutes data.
}
