% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plotchart_aws.R
\name{displayTableAgrrDataSel}
\alias{displayTableAgrrDataSel}
\title{Get aggregated data.}
\usage{
displayTableAgrrDataSel(tstep, aws, vars, pars, start, end, aws_dir)
}
\arguments{
\item{tstep}{time step.}

\item{aws}{vector of AWS IDs.}

\item{vars}{variables.}

\item{pars}{parameters.}

\item{start}{start time.}

\item{end}{end time.}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get aggregated data to display on table for multiple AWS.
}
