% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plotmap_mslp.R
\name{dispMapMSLPHourly}
\alias{dispMapMSLPHourly}
\title{Compute hourly mean sea level pressure.}
\usage{
dispMapMSLPHourly(time, aws_dir)
}
\arguments{
\item{time}{the time to display in the format "YYYY-MM-DD-HH"}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Compute hourly mean sea level pressure data to display on map.
}
