% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_display_winddata.R
\name{dispWindBarb}
\alias{dispWindBarb}
\title{Get wind data.}
\usage{
dispWindBarb(tstep, aws, start, end, aws_dir)
}
\arguments{
\item{tstep}{time step of the data.}

\item{aws}{AWS ID.}

\item{start}{start time.}

\item{end}{end time.}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get wind data for wind barb display.
}
