% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_display_winddata.R
\name{dispWindContours}
\alias{dispWindContours}
\title{Plot wind contour.}
\usage{
dispWindContours(tstep, aws, start, end, centre, aws_dir)
}
\arguments{
\item{tstep}{time step of the data.}

\item{aws}{AWS ID.}

\item{start}{start time.}

\item{end}{end time.}

\item{centre}{center of the image, N, W, S, E.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
CSV object
}
\description{
Plot wind contour.
}
