% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{donwWindFreqCSV}
\alias{donwWindFreqCSV}
\title{Get wind frequency.}
\usage{
donwWindFreqCSV(tstep, aws, start, end, aws_dir)
}
\arguments{
\item{tstep}{time step of the data.}

\item{aws}{AWS ID.}

\item{start}{start time.}

\item{end}{end time.}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
CSV object
}
\description{
Get wind frequency for download.
}
