% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downAWS10MinSpCSV}
\alias{downAWS10MinSpCSV}
\title{Get AWS 10 minutes spatial data.}
\usage{
downAWS10MinSpCSV(time, aws_dir)
}
\arguments{
\item{time}{the time to download in the format "YYYY-MM-DD-HH-MM".}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
CSV object
}
\description{
Get AWS 10 minutes spatial data for download.
}
