% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downAWSAggrCDTDataCSV}
\alias{downAWSAggrCDTDataCSV}
\title{Get aggregated data in CDT format.}
\usage{
downAWSAggrCDTDataCSV(tstep, vars, pars, start, end, aws_dir)
}
\arguments{
\item{tstep}{time step.}

\item{vars}{variable.}

\item{pars}{parameter.}

\item{start}{start date.}

\item{end}{end date.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
CSV object
}
\description{
Get aggregated data in CDT format for download.
}
