% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downAWSAggrOneVarCSV}
\alias{downAWSAggrOneVarCSV}
\title{Get aggregated data.}
\usage{
downAWSAggrOneVarCSV(tstep, aws, vars, start, end, aws_net, aws_dir)
}
\arguments{
\item{tstep}{time step.}

\item{aws}{AWS ID.}

\item{vars}{variable.}

\item{start}{start date.}

\item{end}{end date.}

\item{aws_net}{AWS network.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
CSV format object
}
\description{
Get aggregated data displayed on the chart for download.
}
