% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downAWSAggrSpCSV}
\alias{downAWSAggrSpCSV}
\title{Get AWS aggregated spatial data.}
\usage{
downAWSAggrSpCSV(tstep, time, aws_dir)
}
\arguments{
\item{tstep}{the time step of the data.}

\item{time}{the time to display in the format,
hourly: "YYYY-MM-DD-HH",
daily: "YYYY-MM-DD",
pentad: "YYYY-MM-DD",
dekadal: "YYYY-MM-DD",
monthly: "YYYY-MM"}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
CSV object
}
\description{
Get AWS aggregated spatial data for download.
}
