% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downAWSMinDataCSV}
\alias{downAWSMinDataCSV}
\title{Get 10 minutes data.}
\usage{
downAWSMinDataCSV(aws, vars, start, end, aws_net, aws_dir)
}
\arguments{
\item{aws}{AWS ID.}

\item{vars}{variables.}

\item{start}{start time.}

\item{end}{end time.}

\item{aws_net}{AWS network.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
CSV format object
}
\description{
Get 10 minutes data for download.
}
