% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downMSLPHourly}
\alias{downMSLPHourly}
\title{Compute hourly mean sea level pressure.}
\usage{
downMSLPHourly(time, aws_dir)
}
\arguments{
\item{time}{target date.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
CSV object
}
\description{
Compute hourly mean sea level pressure
}
