% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downRainAccumulSP}
\alias{downRainAccumulSP}
\title{Compute precipitation accumulation.}
\usage{
downRainAccumulSP(tstep, time, accumul, aws_dir)
}
\arguments{
\item{tstep}{time basis to accumulate the data.}

\item{time}{target date.}

\item{accumul}{accumulation duration.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
CSV object
}
\description{
Compute precipitation accumulation for download.
}
