% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_awsdata.R
\name{downTableAggrDataSelCSV}
\alias{downTableAggrDataSelCSV}
\title{Get aggregated data.}
\usage{
downTableAggrDataSelCSV(tstep, aws, vars, pars, start, end, aws_dir)
}
\arguments{
\item{tstep}{time step.}

\item{aws}{vector of AWS IDs.}

\item{vars}{variables.}

\item{pars}{parameters.}

\item{start}{start time.}

\item{end}{end time.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
CSV format object
}
\description{
Get aggregated data to download for multiple AWS.
}
