% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_limit_check.R
\name{qc_limit_check}
\alias{qc_limit_check}
\title{Perform limit check.}
\usage{
qc_limit_check(dirAWS, netAWS)
}
\arguments{
\item{dirAWS}{full path to the directory of the parsed data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}

\item{netAWS}{the name of the AWS network, "LSI-ELOG", "LSI-XLOG" or "REMA".}
}
\description{
Perform quality control using limit check method at AWS time step.
}
