% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_limit_check.R
\name{qc_limit_check_arch}
\alias{qc_limit_check_arch}
\title{Perform limit check archive mode.}
\usage{
qc_limit_check_arch(start_time, end_time, dirAWS, netAWS)
}
\arguments{
\item{start_time}{the start time to process in the format "YYYY-MM-DD HH:MM".
Example: "2019-12-15 12:50"}

\item{end_time}{the end time to process in the format "YYYY-MM-DD HH:MM"}

\item{dirAWS}{full path to the directory of the parsed data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}

\item{netAWS}{the name of the AWS network, "LSI-ELOG", "LSI-XLOG" or "REMA".}
}
\description{
Perform quality control using limit check method at AWS time step.
}
