% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_qc_spatial_check.R
\name{qc_spatial_check}
\alias{qc_spatial_check}
\title{Perform spatial check for hourly data.}
\usage{
qc_spatial_check(start_time, end_time, dirAWS)
}
\arguments{
\item{start_time}{the start time to process, format "YYYY-MM-DD HH:00".
Example: "2019-12-15 12:00"}

\item{end_time}{the end time to process, format "YYYY-MM-DD HH:00"}

\item{dirAWS}{full path to the directory of the parsed data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\description{
Perform quality control using spatial check method at hourly time step.
}
