% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{readAWSStatus}
\alias{readAWSStatus}
\title{Get AWS Status data.}
\usage{
readAWSStatus(hour, aws_dir)
}
\arguments{
\item{hour}{an integer, the last hour to display.}

\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get AWS Status data to display on map.
}
