% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{readAWSWind}
\alias{readAWSWind}
\title{Get AWS Wind coordinates.}
\usage{
readAWSWind(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get AWS Wind coordinates to display on map.
}
