% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{readCoordsMap}
\alias{readCoordsMap}
\title{Get AWS coordinates.}
\usage{
readCoordsMap(aws_dir)
}
\arguments{
\item{aws_dir}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get AWS coordinates to display on map.
}
