% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_infos_coords.R
\name{readAWSStatus}
\alias{readAWSStatus}
\title{Get AWS Status data.}
\usage{
readAWSStatus(ltime, aws_dir)
}
\arguments{
\item{ltime}{character, the last time duration to display. Options are, "01h", "03h", "06h", 
"12h", "24h", "02d", "03d", "05d", "01w", "02w", "03w", "01m".}

\item{aws_dir}{full path to the directory of the AWS data.\cr
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\value{
a JSON object
}
\description{
Get AWS Status data to display on map.
}
