% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_aws_status.R
\name{update_aws_status}
\alias{update_aws_status}
\title{Update AWS data availability.}
\usage{
update_aws_status(dirAWS)
}
\arguments{
\item{dirAWS}{full path to the directory of the AWS data.
Example: "/home/data/MeteoRwanda_Data/AWS_DATA"}
}
\description{
Update AWS data availability every hour.
}
